A pass at changing how constructors work.
This changes the instance construction syntax from ``cls()`` to ``cls.new()``.
The reason this change is made is that there are a lot of circumstances in
which you'll want your constructor to return something other than just ``cls``,
``Option<cls>`` for example is expected to be common. Therefore you don't to
force it to return only ``cls``, and ``cls()`` returning something else seems
weird.

Points for discussion:

* ``classdef`` syntax is kind of awful
* Does this distinction make sense?
* Is there some other way to accomplish it?

Rust uses something similar to this approach, Go's approach is that ``cls{}``
is a "data only" constructor, and types which wish to provide alternate
constructors can provide module nevel ``NewCls`` functions.

The idea of having ``cls(...)`` as a private constructor, and then encourage
people to expose other, public constructors as class methods is also a
possibility.

Feedback?